requirements.txt
Flask==2.0.1
pycryptodome==3.10.1
Werkzeug==2.0.1 